some cleanup, fix aoe breaking speed, initial work on tree felling
