+++ src/main/java/gregtech/common/items/tool/HarvestIceBehavior.java
@@ -1,7 +1,7 @@
 package gregtech.common.items.tool;
 
 import gregtech.api.items.toolitem.ToolHelper;
-import gregtech.api.items.toolitem.behaviour.IToolBehavior;
+import gregtech.api.items.toolitem.behavior.IToolBehavior;
 import net.minecraft.client.resources.I18n;
 import net.minecraft.client.util.ITooltipFlag;
 import net.minecraft.item.ItemStack;
@@ -23,6 +23,6 @@ public class HarvestIceBehavior implements IToolBehavior {
 
     @Override
     public void addInformation(@Nonnull ItemStack stack, @Nullable World world, @Nonnull List<String> tooltip, @Nonnull ITooltipFlag flag) {
-        tooltip.add(I18n.format("metaitem.tool.behavior.silk_ice"));
+        tooltip.add(" " + I18n.format("item.gt.tool.behavior.silk_ice"));
     }
 }
