+++ src/main/java/gregtech/common/items/tool/TreeFellingBehavior.java
@@ -4,29 +4,31 @@ import gregtech.api.items.toolitem.ToolHelper;
 import gregtech.api.items.toolitem.behavior.IToolBehavior;
 import net.minecraft.client.resources.I18n;
 import net.minecraft.client.util.ITooltipFlag;
-import net.minecraft.entity.EntityLivingBase;
+import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
+import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.util.List;
 
-public class DisableShieldBehavior implements IToolBehavior {
-
-    @Override
-    public boolean canDisableShield(ItemStack stack, ItemStack shield, EntityLivingBase entity, EntityLivingBase attacker) {
-        return true;
-    }
+/**
+ * The Tree Felling Behavior must be handled in a special way in {@link gregtech.api.items.toolitem.IGTTool#definition$onBlockStartBreak(ItemStack, BlockPos, EntityPlayer)}
+ *
+ * @see gregtech.api.items.toolitem.ToolHelper#treeFellingRoutine(EntityPlayerMP, ItemStack, BlockPos)
+ */
+public class TreeFellingBehavior implements IToolBehavior {
 
     @Override
     public void addInformation(@Nonnull ItemStack stack, @Nullable World world, @Nonnull List<String> tooltip, @Nonnull ITooltipFlag flag) {
-        tooltip.add(" " + I18n.format("item.gt.tool.behavior.shield_disable"));
+        tooltip.add(" " + I18n.format("item.gt.tool.behavior.torch_place"));
     }
 
     @Override
     public void addBehaviorNBT(@Nonnull ItemStack stack, @Nonnull NBTTagCompound tag) {
-        tag.setBoolean(ToolHelper.DISABLE_SHIELDS_KEY, true);
+        tag.setBoolean(ToolHelper.TREE_FELLING_KEY, true);
     }
 }
