--- src/main/resources/assets/gregtech/models/item/material_sets/dull/tool_head_plow.json
@@ -1,6 +1,0 @@
-{
-  "parent": "item/generated",
-  "textures": {
-    "layer0": "gregtech:items/material_sets/dull/tool_head_plow"
-  }
-}
