--- src/main/resources/assets/gregtech/models/item/material_sets/magnetic/tool_head_plow.json
@@ -1,6 +1,0 @@
-{
-  "parent": "gregtech:item/material_sets/dull/tool_head_plow",
-  "textures": {
-    "layer1": "gregtech:items/material_sets/magnetic/magnetic_overlay"
-  }
-}
