--- src/main/resources/assets/gregtech/models/item/metaitems/jackhammer_base.json
@@ -1,6 +1,0 @@
-{
-  "parent": "item/generated",
-  "textures": {
-    "layer0": "gregtech:items/tools/jackhammer_base"
-  }
-}
