--- src/main/resources/assets/gregtech/models/item/tools/jackhammer.json
@@ -1,7 +1,0 @@
-{
-  "parent": "item/handheld",
-  "textures": {
-    "layer0": "gregtech:items/tools/jackhammer_base",
-    "layer1": "gregtech:items/tools/jackhammer"
-  }
-}
