--- src/main/resources/assets/gregtech/models/item/tools/plow.json
@@ -1,7 +1,0 @@
-{
-  "parent": "item/handheld",
-  "textures": {
-    "layer0": "gregtech:items/tools/handle",
-    "layer1": "gregtech:items/tools/plow"
-  }
-}
