--- src/main/resources/assets/gregtech/models/item/tools/plunger.json
@@ -1,8 +1,0 @@
-{
-  "parent": "item/handheld",
-  "textures": {
-    "layer0": "gregtech:items/void",
-    "layer1": "gregtech:items/tools/plunger_base",
-    "layer2": "gregtech:items/tools/plunger"
-  }
-}
