+++ src/main/resources/assets/gregtech/models/item/tools/wrench_iv.json
@@ -1,7 +1,7 @@
 {
   "parent": "item/handheld",
   "textures": {
-    "layer0": "gregtech:items/tools/power_unit_mv",
+    "layer0": "gregtech:items/tools/power_unit_iv",
     "layer1": "gregtech:items/tools/wrench_electric"
   }
 }
